fix: make filter drawer a common component Fixes #369
