+++ src/app/community/[communityId]/common/filter-component/index.tsx
@@ -1,5 +1,6 @@
 export { EventSelect } from './event-select';
 export { FilterChip } from './filter-chip';
+export * from './filter-drawer';
 export { GpsSelect } from './gps-select';
 export { TicketSelect } from './ticket-select';
 export { YearSelect } from './year-select';
